src/main/java/seedu/unify/model/task/Time.java
package seedu.unify.model.task;

import static java.util.Objects.requireNonNull;
import static seedu.unify.commons.util.AppUtil.checkArgument;

import java.time.LocalTime;

/**
 * Represents a Task's due time in the task list.
 * Guarantees: immutable; is valid as declared in {@link #isValidTime(String)}
 */
public class Time {

    public static final String MESSAGE_CONSTRAINTS = "Time should only contain numbers in the format of HH:MM, "
            + "where HH is from 00 to 23 and MM is from 00 to 59.";
    private static final String HOUR_REGEX = "([01][0-9]|2[0-3])";
    private static final String MINUTE_REGEX = "([0-5][0-9])";
    public static final String VALIDATION_REGEX = HOUR_REGEX + ":" + MINUTE_REGEX;

    public final LocalTime time;

    /**
     * Constructs a {@code Time}.
     *
     * @param time A valid time.
     */
    public Time(String time) {
        requireNonNull(time);
        checkArgument(isValidTime(time), MESSAGE_CONSTRAINTS);
        this.time = LocalTime.parse(time);
    }

    /**
     * Returns true if a given string is a valid time.
     */
    public static boolean isValidTime(String test) {
        return test.matches(VALIDATION_REGEX);
    }

    @Override
    public String toString() {
        return time.toString();
    }

    @Override
    public boolean equals(Object other) {
        return other == this // short circuit if same object
                || (other instanceof Time // instanceof handles nulls
                && time.equals(((Time) other).time)); // state check
    }

    @Override
    public int hashCode() {
        return time.hashCode();
    }

}
